Manuel transcations, single item loading, login loading error

+++ lib/domain/model/transcaton/get_credited_transcations_response_model/transcation.dart
@@ -6,8 +6,8 @@ part 'transcation.g.dart';
 class Transcation {
   String? type;
   String? paymentId;
-  int? price;
-  int? gstPrice;
+  double? price;
+  double? gstPrice;
   int? gstPercentage;
   String? partnerState;
   @JsonKey(name: 'HomeState')
